// packages/components/src/upload/src/interface.ts
export interface FileInfo {
  id: string
  name: string
  batchId?: string | null
  percentage?: number | null
  status: 'pending' | 'uploading' | 'finished' | 'removed' | 'error'
  url?: string | null
  file?: File | null
  thumbnailUrl?: string | null
  type?: string | null
  fullPath?: string | null
}

export type SettledFileInfo = Required<FileInfo>

export type ShouldUseThumbnailUrl = (file: SettledFileInfo) => boolean
// ...
export type OnRemove = (data: {
  file: SettledFileInfo
  fileList: SettledFileInfo[]
}) => Promise<boolean> | boolean | any

export type OnDownload = (
  file: SettledFileInfo
) => Promise<boolean> | boolean | any

export interface UploadInternalInst {
  doChange: DoChange
  xhrMap: Map<string, XMLHttpRequest>
  isErrorState: ((xhr: XMLHttpRequest) => boolean) | undefined
  onError: OnError | undefined
  onFinish: OnFinish | undefined
}

export type DoChange = (
  fileAfterChange: SettledFileInfo,
  event?: ProgressEvent | Event,
  options?: {
    append?: boolean
    remove?: boolean
  }
) => void

export type OnUpdateFileList = (fileList: SettledFileInfo[]) => void

export type RenderIcon = (file: SettledFileInfo) => VNodeChild

export interface UploadInjection {
  mergedClsPrefixRef: Ref<string>
  mergedThemeRef: Ref<MergedTheme<UploadTheme>>
  showCancelButtonRef: Ref<boolean>
  showRemoveButtonRef: Ref<boolean>
  showDownloadButtonRef: Ref<boolean>
  showRetryButtonRef: Ref<boolean>
  showTriggerRef: Ref<boolean>
  mergedFileListRef: Ref<SettledFileInfo[]>
  onRemoveRef: Ref<OnRemove | undefined>
  onDownloadRef: Ref<OnDownload | undefined>
  xhrMap: Map<string, XMLHttpRequest>
  showPreviewButtonRef: Ref<boolean>
  onPreviewRef: Ref<OnPreview | undefined>
  listTypeRef: Ref<ListType>
  dragOverRef: Ref<boolean>
  draggerInsideRef: { value: boolean }
  fileListStyleRef: Ref<string | CSSProperties | undefined>
  mergedDisabledRef: Ref<boolean>
  maxReachedRef: Ref<boolean>
  abstractRef: Ref<boolean>
  imageGroupPropsRef: Ref<ImageGroupProps | undefined>
  cssVarsRef: undefined | Ref<CSSProperties>
  themeClassRef: undefined | Ref<string>
  mergedDirectoryDndRef: Ref<boolean>
  acceptRef: Ref<string | undefined>
  triggerStyleRef: Ref<CSSProperties | string | undefined>
  doChange: DoChange
  onRender: undefined | (() => void)
  submit: (fileId?: string) => void
  shouldUseThumbnailUrlRef: Ref<ShouldUseThumbnailUrl>
  getFileThumbnailUrlResolver: (
    file: SettledFileInfo
  ) => Promise<string> | string
  renderIconRef: Ref<RenderIcon | undefined>
  handleFileAddition: (files: FileAndEntry[] | null, e?: Event) => void
  openOpenFileDialog: () => void
}

export const uploadInjectionKey =
  createInjectionKey<UploadInjection>('z-upload')
// ...
export type ListType = 'text' | 'image' | 'image-card'

export type OnPreview = (file: SettledFileInfo) => void
// ...
export interface FileAndEntry {
  file: File
  entry: FileSystemFileEntry | null
  source: 'dnd' | 'input'
}

export type IconFunction = () => JSX.Element

type RenderFn<T> = () => T

export interface ActionButtonProps {
  key: string
  clickFn: (e: MouseEvent) => void
  slotValue:
  | JSX.Element
  | string
  | { icon: IconFunction }
  | RenderFn<string>
  | RenderFn<JSX.Element>
  isCircle?: boolean
  size?: Size
}